refactor: extract ConflictResolver from SftpDataAdapter (#199)
Move the 3-way merge and two-choice conflict resolution logic into a
dedicated `ConflictResolver` class in `conflict/ConflictResolver.ts`.

SftpDataAdapter's constructor replaces three conflict-related params
(onWriteConflict, ancestorTracker callback, onTextConflict) with a
single `conflictResolver: ConflictResolver | null`, reducing the file
from 1003 to ~920 lines.

Closes #199
